+++ src/lib/client/graphql/mod.rs
@@ -2,11 +2,11 @@ mod pr_title;
 mod prs;
 
 pub(crate) use pr_title::get_pull_request_title;
-pub(crate) use prs::Edge;
+use prs::PrItem;
 
 use crate::Error;
 
 pub(crate) trait GraphQL {
     #[allow(async_fn_in_trait)]
-    async fn get_open_pull_requests(&self) -> Result<Vec<Edge>, Error>;
+    async fn get_open_pull_requests(&self) -> Result<Vec<PrItem>, Error>;
 }