+++ CHANGELOG.md
@@ -12,12 +12,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - use commit and push logic for changelog update(pr [#301])
 - add Rebase command to CLI(pr [#304])
 - add graphql method to get open pull requests(pr [#306])
+<!-- markdownlint-disable MD033 -->
+- add PrItem struct and return Vec<PrItem> instead of Vec<Edge>(pr [#307])
+<!-- markdownlint-enable MD033 -->
 
 ### Changed
 
 - chore-remove commented out debug logs and unused code(pr [#302])
 - remove-remove customised commit and push for changelogs(pr [#303])
-- ci-integrate rebase commanmd into change update job(pr [#305])
+- ci-integrate rebase command into change update job(pr [#305])
 
 ## [0.4.1] - 2024-08-24
 
@@ -585,6 +588,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#304]: https://github.com/jerus-org/pcu/pull/304
 [#305]: https://github.com/jerus-org/pcu/pull/305
 [#306]: https://github.com/jerus-org/pcu/pull/306
+[#307]: https://github.com/jerus-org/pcu/pull/307
 [Unreleased]: https://github.com/jerus-org/pcu/compare/v0.4.1...HEAD
 [0.4.1]: https://github.com/jerus-org/pcu/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/jerus-org/pcu/compare/v0.3.0...v0.4.0